add custom hooks useTodos.ts & useErrorMessage.ts
