+++ src/App.tsx
@@ -1,22 +1,22 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { UserWarning } from './UserWarning';
-import { createTodo, deleteTodo, USER_ID } from './api/todos';
-import { Todo } from './types/Todo';
-import { client } from './utils/fetchClient';
+import { USER_ID } from './api/todos';
 import { TodoList } from './components/TodoList';
 import { Header } from './components/Header';
 import { Footer } from './components/Footer';
 import { ErrorNotification } from './components/ErrorNotificaton';
 import { FilterStatus } from './types/types';
+import { useErrorMessage } from './hooks/useErrorMessage';
+import { useTodos } from './hooks/useTodos';
 
 export const App: React.FC = () => {
-  const [todos, setTodos] = useState<Todo[]>([]);
-  const [tempTodo, setTempTodo] = useState<Todo | null>(null);
   const [filter, setFilter] = useState<FilterStatus>(FilterStatus.All);
-  const [error, setError] = useState<string | null>(null);
-  const [deletingId, setDeletingId] = useState<number[]>([]);
+  const [error, setError] = useErrorMessage();
+
+  const { todos, tempTodo, deletingId, addTodo, removeTodo, clearCompleted } =
+    useTodos(setError);
 
   const visibleTodos = todos.filter(todo => {
     if (filter === FilterStatus.Active) {
@@ -30,72 +30,9 @@ export const App: React.FC = () => {
     return true;
   });
 
-  const addTodo = async (title: string) => {
-    const trimmedTitle = title.trim();
-
-    setTempTodo({
-      id: 0,
-      title: trimmedTitle,
-      completed: false,
-      userId: USER_ID,
-    });
-
-    try {
-      const newTodo = await createTodo({
-        title: trimmedTitle,
-        completed: false,
-        userId: USER_ID,
-      });
-
-      setTodos(prev => [...prev, newTodo]);
-    } catch {
-      setError('Unable to add a todo');
-      throw new Error();
-    } finally {
-      setTempTodo(null);
-    }
-  };
-
-  const removeTodo = async (todoId: number) => {
-    setDeletingId(prev => [...prev, todoId]);
-    setError(null);
-
-    try {
-      await deleteTodo(todoId);
-      setTodos(prev => prev.filter(todo => todo.id !== todoId));
-    } catch {
-      setError('Unable to delete a todo');
-    } finally {
-      setDeletingId(prev => prev.filter(id => id !== todoId));
-    }
-  };
-
-  const clearCompleted = async () => {
-    const completedTodos = todos.filter(todo => todo.completed);
-
-    await Promise.all(completedTodos.map(todo => removeTodo(todo.id)));
-  };
-
-  useEffect(() => {
-    client
-      .get<Todo[]>(`/todos?userId=${USER_ID}`)
-      .then(setTodos)
-      .catch(() => setError('Unable to load todos'));
-  }, []);
-
-  useEffect(() => {
-    if (error) {
-      const timer = setTimeout(() => setError(null), 3000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [error]);
-
   if (!USER_ID) {
     return <UserWarning />;
   }
-
-  // console.log(visibleTodos.length);
 
   return (
     <div className="todoapp">
